# Use `f32` arithmetic for `f32` normal distribution