--- a/rand_distr/src/normal.rs
+++ b/rand_distr/src/normal.rs
@@ -481,11 +481,11 @@ mod tests {
 
         let e = core::f32::consts::E;
         let lnorm = LogNormal::from_mean_cv(e.sqrt(), (e - 1.0).sqrt()).unwrap();
-        assert_almost_eq!(lnorm.norm.mean, 0.0, 2e-16);
-        assert_almost_eq!(lnorm.norm.std_dev, 1.0, 2e-16);
+        assert_almost_eq!(lnorm.norm.mean, 0.0, 2e-7);
+        assert_almost_eq!(lnorm.norm.std_dev, 1.0, 2e-7);
 
         let lnorm = LogNormal::from_mean_cv(e.powf(1.5), (e - 1.0).sqrt()).unwrap();
-        assert_almost_eq!(lnorm.norm.mean, 1.0, 1e-15);
+        assert_almost_eq!(lnorm.norm.mean, 1.0, 1e-6);
         assert_eq!(lnorm.norm.std_dev, 1.0);
     }