--- a/rand_distr/src/normal.rs
+++ b/rand_distr/src/normal.rs
@@ -486,7 +486,7 @@ mod tests {
 
         let lnorm = LogNormal::from_mean_cv(e.powf(1.5), (e - 1.0).sqrt()).unwrap();
         assert_almost_eq!(lnorm.norm.mean, 1.0, 1e-6);
-        assert_eq!(lnorm.norm.std_dev, 1.0);
+        assert_almost_eq!(lnorm.norm.std_dev, 1.0, 1e-6);
     }
 
     #[test]