# Summary

The current normal distribution `f32` implementation just computes the `f64` version and converts the result to `f32`.

- [x] Added a `CHANGELOG.md` entry

# Motivation

The TODO suggested this, and I agreed. This should be significantly faster on platforms where single-precision floats are faster.

# Details

I converted the code to use `f32` arithmetic and adjusted the number of bits used for the calculation of `u`, the base number.

In addition to the tests, I also generated some normal numbers with the current f64 and the new f32 paths and ensured that they both look "normal".

Tests that depended on the specific `f32` values generated by the old normal distribution code had to be updated, since we now use `u32`-sized random numbers as the source, which will be different from the `u64`s used previously.

I also fixed some comments that were confusing and possibly incorrect in the ziggurat code.

I was originally tempted to convert the ziggurat tables to `f32` for all versions, but this appears to affect the precision of the `f64` generated numbers too much.