@swenson 2024-08-09
Hmm. Benchmarks for this are not encouraging on my Threadripper 3970X (Zen 2), which I have heard are particularly weak in their FPU.

Before:
```
normal/normal           time:   [11874.0815 cycles 11893.9016 cycles 11912.4901 cycles]
                        thrpt:  [2.9781 cpb 2.9735 cpb 2.9685 cpb]
```

After (with this change):
```
normal/normal           time:   [19646.2155 cycles 19662.4413 cycles 19682.2019 cycles]
                        thrpt:  [4.9206 cpb 4.9156 cpb 4.9116 cpb]
                 change:
                        time:   [+65.022% +65.173% +65.332%] (p = 0.00 < 0.05)
                        thrpt:  [-39.516% -39.457% -39.402%]
```

I'll try to test on an Intel CPU later to see if it is different. (I was not able to compile on my arm64 due to `criterion`.)