@swenson 2024-08-09
Intel (Core Ultra 7 165H) before:
```
normal/normal           time:   [9380.9673 cycles 9662.9064 cycles 9959.9462 cycles]
                        thrpt:  [2.4900 cpb 2.4157 cpb 2.3452 cpb]
```

After:
```
normal/normal           time:   [13374.5863 cycles 13813.0605 cycles 14315.2989 cycles]
                        thrpt:  [3.5788 cpb 3.4533 cpb 3.3436 cpb]
                 change:
                        time:   [+34.629% +41.460% +48.790%] (p = 0.00 < 0.05)
                        thrpt:  [-32.791% -29.309% -25.722%]
```

Better, but still a serious regression for `f32` performance on modern x86-64 platforms.

Which makes me think it is not worth doing this, or it we should lock it behind a `feature` or architecture-specific gating?