@dhardy 2024-08-10
If someone has a use-case (a platform where this change is beneficial), I'd like to see their benchmarks/evidence here.

Complication: this is a value-breaking change. This is an argument against implementing this via a feature-flag (though not a complete blocker).

I suggest we leave this open for a couple of weeks to see if anyone can provide evidence of benefit, otherwise close (we can revisit later if and when evidence pops up).