@swenson 2024-08-10
Makes sense.

It is possible to make it less value-breaking if we generated `u64` random values and pulled bits in a similar way to the `f64` code path, though the results might differ slightly due to rounding. But, that is potentially slower and probably still won't be exactly the same.

I have a bunch of 32-bit microcontroller development boards, but none that run a real OS, so I'd have to write a custom benchmarking program for them to see if this is beneficial.