@dhardy 2024-08-12
As you say, results would still not be identical so I don't see the benefit of that approach.

It's unclear to me if this is something that would benefit you personally or just some TODO you found. If the former, please quantify.