@swenson 2024-08-12
This was mostly a TODO I found that would be simple enough and possibly worth the optimization. (I do a lot of 32-bit embedded development, but rarely use FP.)